+++ core-application/src/main/java/gohigher/port/in/OAuth2CommandPort.java
@@ -1,7 +1,7 @@
 package gohigher.port.in;
 
+import gohigher.user.Provider;
 import gohigher.user.User;
-import gohigher.user.oauth2.Provider;
 
 public interface OAuth2CommandPort {
