+++ core-domain/src/main/java/gohigher/user/Provider.java
@@ -1,4 +1,4 @@
-package gohigher.user.oauth2;
+package gohigher.user;
 
 import java.util.Locale;
