+++ core-domain/src/main/java/gohigher/user/User.java
@@ -1,6 +1,5 @@
 package gohigher.user;
 
-import gohigher.user.oauth2.Provider;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
