+++ in-adapter-api/src/main/java/gohigher/oauth2/user/OAuth2UserInfoFactory.java
@@ -2,7 +2,7 @@ package gohigher.oauth2.user;
 
 import java.util.Map;
 
-import gohigher.user.oauth2.Provider;
+import gohigher.user.Provider;
 
 public class OAuth2UserInfoFactory {
